requirements.txt
RPi.GPIO
smbus
paho-mqtt
asyncio